Fix more circular import errors
