+++ breadbox/tests/service/test_slice.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 
 from breadbox.db.session import SessionWithUser
-from breadbox.crud.slice import get_slice_data
+from breadbox.service.slice import get_slice_data
 from breadbox.models.dataset import AnnotationType
 from breadbox.schemas.dataset import ColumnMetadata
